RAG/RAG/requirements.txt
transformers==4.40.0
sentence-transformers
faiss-cpu
pypdf 